requirements.txt
requests
beautifulsoup4
pandas
numpy
selenium
webdriver-manager
scrapy
lxml
html5lib
nltk
spacy
sentence-transformers
faiss-cpu
chromadb
openai
python-dotenv
streamlit
plotly
scikit-learn
transformers
torch
langchain
langchain-openai
langchain-chroma
langchain-community
langdetect
google-generativeai
mistralai 